Added a test (addrNoImplSEI)
This test ensures that an service implementation does not
need to implement the SEI (Service Endpoint Interface).

An test.addrNoImplSEI.AddressBookSoapImpl is added that
is just like the samples.addr.AddressBookSoapImpl except that
it does not implement AddressBook.

I also removed the <operation> elements from the
deploy.wsdd for this test to force the runtime to use introspection.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247088 13f79535-47bb-0310-9956-ffa450edef68