I've added a bit of JAX-RPC implementation.  JAX-RPC says getPort can return either a generated stub or a dynamic proxy.  We just returned the dyamic proxy.  But sometimes we have enough info to return a generated stub, if it's available.  So now we try to return the generated stub where we can and, if it fails, only then return the dynamic proxy.
This change needed a utility-like method that was in JavaServiceImplWriter:
getAddressFromPort.  This method doesn't quite fit into any utility class we
have, so I created a new one - WSDLUtils - for WSDL4J utility methods.  A
fair bit of what's in the Java writers COULD be put into this class, but for
now getAddressFromPort is the only thing in there.

I've also written a simple test to make sure the generated stub is returned
when it's available.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254174 13f79535-47bb-0310-9956-ffa450edef68