Patch from John Brothers [johnbr@undefined.com]:
I have been learning SOAP via Axis, and I encountered a
NullPointerException in RPCElement() when trying to test
out a new service that I had written.  I eventually discovered
that this was caused by Axis's inability to find the class file(s)
for my service (I had used a link instead of copying the directory).

I submit that it would be more informative to the user
(and more elegant) to throw a ClassNotFoundException instead
of a NullPointerException in that case, to help the user more
quickly discover the "directory tree copy" requirement.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247178 13f79535-47bb-0310-9956-ffa450edef68