Simplify, simplify...
Redo the fix dims made, but without the extra class.  The problem was that
SOAPHandler wasn't pushing the associated MessageElement during
startElement() if it existed prior to the startElement() call.  Now we always
do it, which also eliminated the need to manually push the BodyElement
in BodyBuilder.onStartChild().

Eventually, this mechanism (pushing new MessageElements) might move
entirely out of the SOAPHandler's realm of responsibility, so that the
framework (DeserializaitonContextImpl) just deals with it.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247324 13f79535-47bb-0310-9956-ffa450edef68