The DefaultTypeMappingImpl contains the mappings for wsdl <-> java that are compliant with JAXRPC
The DefaultSOAP12TypeMappingImpl (which is inappropriately named) is the
same as the DefaultTypeMappingImpl except that the soap-enc mappings are
preferred over xsd mappings.

Both WSDL2Java and Java2WSDL have options which allow either of these
mappings to be chosen.  (1.1 gets the first one, 1.2 gets the second one).

-------------------------------------------------------
The default for these options was 1.2.  The default is changed to 1.1 for compliance.
-------------------------------------------------------

Further notes:
Glen, pick up this change if you want it in b3.

After things settle, we need to have a better solution going forward.
The following options seems more reasonable:
   1.1, 1.1encoded, 1.2, 1.2encoded

We can discuss these later.



PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254537 13f79535-47bb-0310-9956-ffa450edef68