The following new options are added to Java2WSDL   -I, --input <wsdl>   -b, --bindingName <name>
The --input <wsdl> option is used to read an existing
wsdl file and use it as the javax.wsdl.Definition.

The code is changed in several places to query for
an existing wsdl construct before adding a new one.
For example a binding is created only if a
binding with the same qname does not exist.

Ignoring existing constructs is a simple rule.  (An alternative
rule would be to replace the existing construct, but
a replacement rule is more error prone and less desirable.)

Dealing with the types was a bit more complicated due to the
way types are processed.  The Types class builds a SymbolTable
with the input wsdl file and examines the types in the symbol table
to determine which types/elements do not need to be reprocessed.
This has the beneficial side effect of checking the
input wsdl file.

The new code exposed some minor bugs, which I fixed.

---------------------

The --binding <name> option is used to set the name of the
wsdl binding element.

---------------------
I used these new options to generate wsdl for the TCK
tests that required multi-ports, mult-services, multi-portTypes, etc.

The --input <wsdl> option will also be useful in the future to
add such things as doc-lit bindings.  (First run the tool to generate
a wsdl with the RPC binding, and then run again with the --input <wsdl>
to add a doc-lit binding).

--------------------------------------------

I also changed some Java2WSDL signatures to be
more specific in regards to Exceptions.

--------------------------------------------
I expect that there will be additions to this code as it matures.
I need to add information to the user guide and add a testcase.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254585 13f79535-47bb-0310-9956-ffa450edef68