If we have
<message name="in"><part name="in" element="e"/></message>
<message name="out"><part name="out" element="e"/></message>

where e contains a string named a.  Note that the input and output
messages both refer to the same element.  Because the unwrapped list
of parameters are named the same for input and output, we used to say
we had an inout parameter and generate the following:

public void e(StringHolder a);

but inoutness is determined by part name, not parameter name, so we
should generate:

public String e(String a);

This is a TCK issue.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247514 13f79535-47bb-0310-9956-ffa450edef68