Fix for Bugzilla Bug 10646: When echo?wsdl is requested, the generated wsdl uses xsd:base64Binary to describe the echoHexBinary request and response.
Solution:

The root of this problem is line 1070 in the Java2WSDL Emitter.
The writePartType takes an (javaType, xmlType) pair to write the
message part.  The Emitter was sending the proper javaType, but
was sending the name of the element instead of the qname representing
the type.  The wrong combination of arguments meant that writePartType
always resorted to intropspecting javaType to determine the
type to write.  And in this case, a byte[] maps to xsd:base64binary.

Changing the Emitter code to pass in the correct QName fixed the 10646
bug, but revealed a number of other hidden bugs during testing.

The following additional changes were made:

  * Changed the Types:writeType(javaType) method to
    Types:writeType(javaType, qName) so that it could be called
    with both the javaType and xmlType pairing.  Similar change
    to writeTypeNamespace.

  * Removed defaulting to the DefaultSOAP12TypeMapping.

  * Fixed some problems with the setting up the TypeMapping on
    the temporary ServiceDesc's set up by the Emitter.

  * Change the java2wsdl calls for the Roundtrip testcase to
    use the appropriate typeMappingVersion.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247553 13f79535-47bb-0310-9956-ffa450edef68