A little more plumbing for SOAP 1.2 support - get the role attribute (actor for 1.1, role for 1.2) from the SOAPConstants.
This really brings to light the next problem which needs solving, namely
how to deal with getting the right SOAPConstants when not in the midst
of a deserialization.  My current suggestion is to have an equivalent
to MessageContext.getCurrentContext() in
SOAPConstants.getCurrentVersion().  Then you can set it for your thread
and have all the classes use it.   Alternately, we could also try to pull
all the soap-version-specific stuff out of the message classes themselves,
and only resolve these things at serialization/deserialization time, when
the appropriate context is guaranteed to be available.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247566 13f79535-47bb-0310-9956-ffa450edef68