Fix problems with the EJBProvider (bug 10864).   http://nagoya.apache.org/bugzilla/show_bug.cgi?id=10864
The getInitializedServiceDesc() function in SOAPService was hard coded to
check the "className" parameter when in fact it should have using the
provider interfaces (which can be overridden) to get the implementation class
of the service for the ServiceDesc.

Added an abstract interface to the BasicProvider class for
getServiceDescription(), which will fill in the implementation class with the
right info.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254695 13f79535-47bb-0310-9956-ffa450edef68