Implement the ServletEndpointContext interface from JAX-RPC.
Create an implementation class in org.apache.axis.transport.http
(Is there a better place for this?)

Create a S.E.C. in the AxisServlet and stash it in the message context
for later.

When we call the init() method of the ServiceLifeCycle, pass the S.E.C
in to the function instead of null.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254707 13f79535-47bb-0310-9956-ffa450edef68