Add support for the XML Schema types gYearMonth, gMonthDay, gYear, gMonth, and gDay.
Remove the Time serializer/deserializer and factories as the SimpleType
classes will handle this type because it has a String constructor and a toString()
method (sweet!).  Used the SimpleType serializer for the new types too.

TODO: The new Unisgned* types can also used this serializer if a
String constructor is added to them.  This will remove a bunch (16!) of
extra encoding.ser classes.

Added Unit tests for the new types and a Deserailization test for each.
TODO: Add g* types to test/wsdl/types test
TODO: We should probably move all the tests for the types in to its own
package and out of test.encoding.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247703 13f79535-47bb-0310-9956-ffa450edef68