This file needed some real work.  First off, there was this problem:
    <!-- NOTE:  THIS DOMAIN HAS EXPIRED!  If it comes back,
         we can add this test back in.
    <wsdl2java url="http://www.perfectxml.net/WebServices/SalesRankNPrice/BookService.asmx?WSDL"
    -->
    <!--
    <wsdl2java url="${axis.home}/test/wsdl/literal/SalesRankNPrice.wsdl"
               output="${axis.home}/build/work"
               verbose="no"
               serverSide="no"
               testcase="no">
        <mapping namespace="http://www.PerfectXML.com/NETWebSvcs/BookService" package="test.wsdl.literal"/>
    </wsdl2java>
    -->

Which meant that nothing was ever generated.  Actually, the comments were the other way around,
with the file commented out first, and a note that the URL didn't exist.  So I flip-flopped
them, and am using the stored wsdl file to generate the classes.

This, in coupling with my last xmls/targets.xml change, now allows this test to succeed.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247733 13f79535-47bb-0310-9956-ffa450edef68