"Minor patches to send XSI type" from "Thomas Sandholm" <sandholm@mcs.anl.gov>
Details from Thomas:
I sent in a mail about this problem earlier. Basically we use a literal
binding and we want to force XSI type not to be sent. However, even if the
SEND_TYPE_ATTR property is turned off in the stubs the serializeActual
method overrides it. The patch I added seems to comply to the commit log
message:

"Change SerializationContext to take a Boolean instead of a boolean
for the "sendTypes" argument.

This allows three values:
true - force sending xsi:type regardless of configuration
false - force no xsi:type regardless of configuration
null - respect configuration (sendXsiTypes option, etc)"


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254828 13f79535-47bb-0310-9956-ffa450edef68