Some more attachments stuff, significantly, Images now work (yay).
But along with Images working, we now have a new dependency:  JIMI.
It's an optional dependency.  If you don't have it AXIS will still build, but if
you want to send Images as attachments, you'll need it.

I don't have an image test, yet, but the TCK tests it, so I'm in no hurry.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254839 13f79535-47bb-0310-9956-ffa450edef68