There must be a way to configure the WSDD service info into the Service Locator object.  The WSDD service's name is typically the same as the WSDL service's port name.  But there is not necessarily a 1-1 correllation between WSDL service's port name and WSDD service name.  A given WSDL service port could have any number of WSDD services.  For instance, imagine a hand-modified client-config.wsdd that has TWO services, one that sends XSI types and one that doesn't.  Right now, one of them would (presumably) be named after the port name and the other would be named "somethingElse".  There is NO WAY to tell the AXIS runtime to use the WSDDService named "somethingElse".
In order to accomplish this feat, I've added to the service Locator class
two methods per port:  get/set<portName>WSDDServiceName.  When the
service locator creates the stub, it passes this info to it.

This brought up one other issue.  Call.setTargetService seemed to be
exactly what I wanted to call, but Service is already calling
Call.setPortName to do what I'm trying to do.  setTargetService and
setPortName appear to be virtually the same thing.  Since setTargetService
wasn't being called and it wasn't documented anywhere, I removed it in
favor of setPortName.  If folks have a problem with this, let me know and I
can return setTargetService and tie it in with setPortName.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254859 13f79535-47bb-0310-9956-ffa450edef68