In the original Xerces from which this originated, SymbolTables were used to convert Strings to integers so that they can be more easily compared.
When transplanted to Axis, this class was used merely to convert Strings
to integers so that they could be converted back again to Strings (upon
which point, the embedded char array is then extracted, appended to a
StringBuffer, extracted as a String, and then converted to the desired
type; but I digress).

More globally, this whole Sax2EventRecorder needs to be optimized away
whenever possible as it is now getting to the point where it is dominating
the processing...

Meanwhile, this commit takes a more modest approach and merely eliminates
the 4 hash gets/put's pair per start element, 3 per end element, and 1 pair
for any character strings encountered in the middle.  This being the
top method on the hit parade on SOAP messages of even moderate size.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247849 13f79535-47bb-0310-9956-ffa450edef68