OOPS!  This file goes along with the new JDK13IO.java!  (commit comments below)
I've discovered a JDK 1.3 implementation for our ImageIO plugin.  The
problem is, it can only write JPEG format, and it's depends on
sun.awt.image.codec.JPEGImageEncoderImpl, though the Sun, the IBM,
and the HP JDKs all have this class.

I've added this class so that those who want to use a non-JIMI JDK1.3
implementation have something they can use, even if it's as limited as this
one is.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254940 13f79535-47bb-0310-9956-ffa450edef68