(forgot a file)
Add a Java Transport/Binding.
This will allow people to switch between native Java calls and SOAP
calls by just switching the WSDL, Port or URL.  The URL is assumed
to be of the form:   java:<classname>    ie. java:org.apache.Foo
This does call all of the client-side chains but since it does not
create a server object it does not call the server side chains.
If you want that then you can use the Local Transport.
This isn't 100% there yet - for example this only handles RPC for now.
Stubbed in code for Msg but i'm running into problems with the Op.Desc.
stuff.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254979 13f79535-47bb-0310-9956-ffa450edef68