Compile the right files (?) when rebuilding this test.  Was failing for me before this, works now.
Can someone explain this test, please?  A readme would be really helpful.

The original WSDL specifies an "address" type (lowercase 'a').  Then
when we Java2WSDL, since the class is called "Address" and there's
no metadata to tell us any different, the new WSDL gets a type
"Address" (uppercase 'A').  So when we test, we're not seeing
wire representations of the original WSDL's messages.  I think this
was the problem before, since it looked like the types weren't
matching, but I'm not exactly sure how the various mixed files
(wsdd + stub) which touch this stuff were being used.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@254985 13f79535-47bb-0310-9956-ffa450edef68