Update from " Greg Truty <gtruty@us.ibm.com>" for "JAX-RPC Handler Support in Axis".
Note from Greg:
Ok... further tests indicate that I had an issue w/my implementation.
Essentially, I am writing a testcase to contribute which add's headers,
checks headers, and modifies the soap body as part of doing this from a
JAX-RPC handler.  It showed that my JAX-RPC intercepters are in the wrong
spot on the server side.  They need to be moved so they run AFTER the pivot
point.  Sending a header in the SOAP response was not possible.  I moved
them to SOAPService and it works fine.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@255017 13f79535-47bb-0310-9956-ffa450edef68