Patch from "Greg Truty" <gtruty@us.ibm.com> for JAXRPC Handler testcase + changes.
Notes from Greg:
I've made a bit of progress w/the Handler code and have testcase plus some
code changes (along w/some other bugs that I have noticed).  This is on a
version of todays driver (9/03).

The testcase is included below in test\jaxrpchandler.  It tests the
addition of a client/server side handler, along w/ client-config and
programatic API access changes.  It deploys/undeploys the service
appropriately.  The handler adds a header element at each of the hook
points, incrementing a counter in the header.  Lastly, on the client
response, it checks to ensure the counter has been updated appropriate and
calls a static method to the testcase (it's out-of-band) to confirm that
the handlers executed appropriately.

The code changes take into account the role/header/parameter options (which
were wrong/missing in the code orginally checked in).


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@247961 13f79535-47bb-0310-9956-ffa450edef68