Fix bug 12833 - Axis clients fail to extract the value of output (only) parameters when they are last in the parameter list
The special case logic I added to prevent functions like
   void echo(StringHolder inout)
and instead make signatures like
   String echo (String in)
was doing very bad things.  Remove it and put the code in a better place
that will not affect processing of functions with more than 1 parameter.

Add a test case which has .NET WSDL for a large number of in/out/inout
and return values, and include the test case noted in the bug for good measure.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@255285 13f79535-47bb-0310-9956-ffa450edef68