Fix for "Must Fix" Bugzilla Defect 12867.
Summary of problem:  Axis Service expected a series of elements in
the maxOccurs="unbounded" form.  A .Net client sent an array containing
the elements instead.  This falls under the "be flexible in what you accept" rule.

Solution: The BeanPropertyDescriptor.set() method can recognize this situation
and set the indexed properties from the items in the deserialized array.

I also added Tests to TestDeser for both situations.
The test exposed a minor flaw in the BeanDeserializer associated with
getting the child deserializers.

All is well.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248219 13f79535-47bb-0310-9956-ffa450edef68