Fix bug 13007 - Java2WSDL emits type=""   http://nagoya.apache.org/bugzilla/show_bug.cgi?id=13007
If we get back a null from writeType() in the BeanSerailizer
type generation code, make the type string xsd:anyType.

NOTE: This will prevent bad WSDL, but it will not allow you to
publish a service with a method that takes anything other than
java.lang.Object.
Example:
  myOp(java.util.Locale inarg)
will not match an operation defined in wsdd with a parameter:
  <parameter name="inarg" type="tns:anyType"
             xmlns:tns="http://www.w3.org/2001/XMLSchema"/>


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@255334 13f79535-47bb-0310-9956-ffa450edef68