Fixes so that WSDL2Java doesn't outright barf on the round 4 group G (SwA/DIME) WSDLs defined at http://www.whitemesa.net/r4/interop4.html
A number of issues remain: for example the WSDDs which are emittted for
mime-rpc.wsdl is devoid of content.  No discernable support exists in
WSDL2Java for DIME.

Given my level of understanding of WSDL2Java (nearly zero), I harbor
no preconceptions that these fixes I have made are other than hacks,
but given the lack of responses I have received to
http://marc.theaimsgroup.com/?l=axis-dev&m=103404539724028&w=2
I felt it worthwhile to check in these changes so that they can be
critiqued and/or built upon.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@255475 13f79535-47bb-0310-9956-ffa450edef68