The recent fault fixes looked, I'm sorry to say, like one big hack.  This isn't particularly surprising considering the mad rush to get faults working before the soap builders' interop meeting.  But it caused me a couple ulcers.
- FaultInfo was an inner class inside JavaDefinitionWriter AND it was used
by SymbolTable.  This was just plain wrong!  There should be absolutely
NO dependencies by the parser/symbol table front end on the emitter back
end.  So I pulled FaultInfo into its own class and put it into the symbolTable
package.

- FaultInfo was nothing but a struct and you had to use Util methods to
massage that info.  Tsk tsk.  Not very OO.  I've moved the logic from
various Utils methods into the FaultInfo class itself.
- The fault changes broke soap:headerfault.  Faults that were declared in
the portType weren't being generated if they appeared in a
soap:headerfault.  They only worked if they appeared in the binding as a
soap:fault.  I've fixed this and uncommented the soap:headerfault bits of
the header test.

A couple remaining problems.
- I still have a serious concern about the fault logic based on the binding
rather than the portType, but that's a bigger fish than I wanted to fry right
now (I'm in the mood for trout and this thing's a whopping flounder).  One
particular issue:  if a WSDL file contains a portType but DOES NOT contain
a binding, fault exceptions would not be generated, therefore, even though
the SEI should be generated, it won't compile.  (There's a more serious
problem in that this sort of WSDL won't even generate the SEI now.  This
broke somewhere along the way.)
- Now that FaultInfo is a real object, it seems like it should really be a
subclass of SymTabEntry, ie., FaultEntry.  But faults aren't first-class
objects in WSDL, so creating a FaultEntry could be problemmatic.  Needs
more thought.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248406 13f79535-47bb-0310-9956-ffa450edef68