Support anonymous types when generating WSDL.
This involved some surgery, because the way it used to be relied on the
fact that the serializers would write out <complexType name=""> types
into the <schema> sections, and there was no way to get your hands on
those elements (to embed them) or make them unnamed.

The big change is that Serializer's writeSchema now returns an Element
and DOES NOT write the complex type into the <schema> itself.  Instead,
we rely on the context in which we're generating the type to know what
to do about it.  This logic is mostly in Types.makeTypeElement(), whose
Javadoc explains how it works.

There was also the usual bits of random cleanup of unused imports, etc.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248562 13f79535-47bb-0310-9956-ffa450edef68