axis-user complaint: the nightly builds wont work on pre1.4 jvms no more.
java1.3 and below had no StringBuffer.append(StringBuffer), but did have StringBuffer.append(Object). So on a 1.3 build this code maps to the Object call, toString() is called and all is well. But on a 1.4 build the new method is used instead, so it aint there at runtime. I wonder how many other places this happens. And I wonder if IDEA has a way to tell me...


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248593 13f79535-47bb-0310-9956-ffa450edef68