using inheritAll to pass down the current env settings to the transport sample.
This sample is special in that by setting the dir param of the <ant> call, we are moving its base dir from axis.home to axis.home/samples/transport, so things that get recreated are relative to a different place. To make matters worse, ant always passes stuff on the command line down, so lib.dir=lib is still there.

Alternate tactics here would be to remove the dir attr and put the complete path in the antfile, or set inheritAll=false and pass down those properties we thing important: axis.home being the key one.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@255894 13f79535-47bb-0310-9956-ffa450edef68