Better defaulting for Call's style / use / encodingStyle attributes.
When you set Style to anything but RPC, we default the Use to LITERAL
(unless Use has been explicitly set).

Whenever either Style is set to RPC or Use is set to ENCODED, we
default the encodingStyle to SOAP-ENC (for whichever version of SOAP
you're using).  Otherwise we default to "" (no encoding).

Got rid of Use.ENCODED12, since nothing was using it, it didn't seem to
do anything, and all the tests pass without it.  Encoded is encoded,
regardless of SOAP version...

Fix add'l SOAP 1.2 tests to use the right namespace for certain args.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248899 13f79535-47bb-0310-9956-ffa450edef68