Better solution to bug #17017.
When we find SOAP arrays (restrictions on SOAP:Array), we always end
up with a DefinedType.  Otherwise, we end up with a CollectionElement,
CollectionType, etc.

So when we have a DefinedType with dimensions (an array), use the
refType (which will correctly be "ArrayOfFoo" in most cases for SOAP
array restrictions).  Otherwise, scan up through the ref types like we used
to to get the "real" XML type.

Passes all tests, and generates the right code for the example attached to
17017.  Also fixes bug:

http://nagoya.apache.org/bugzilla/show_bug.cgi?id=17655

and perhaps also:

http://nagoya.apache.org/bugzilla/show_bug.cgi?id=17705


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@248979 13f79535-47bb-0310-9956-ffa450edef68