this marks all my doc changes for the 1.1 release. Rick Kellogg's 1.0 changes applied, then a review of the install guide.
The big additions there are examples on classpath setup (untested!) and a section on what to expect from that informal support org that is the user mailing list. Harsh but fair, as we like to say. It's less brutal than making the axis download dependent on passing an on-line java/http/webappp competence test, after all


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@249064 13f79535-47bb-0310-9956-ffa450edef68