First docbook addition to the repository, goal would be to migrate all the docs from HTML to .dbk, though not with the app I used to write this. Maybe OpenOffice with DBK support is better.
The subject of the doc is of course some thoughts on testing. I dont come to any compelling answers, but do try and scope out all the ways we really need to test Axis and suggest some changes to what we have today. It should be sufficient to start discussion :)


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@256227 13f79535-47bb-0310-9956-ffa450edef68